Moved function.
